My solution
The java files can be found at /src/lottery/machine/

The solution was created using Net Beans IDE. The entire folder should be opened if executing solution with the same.

LotteryMachine.java
Contains the methods and structures that is the solution to the coding challenge

Test1.java
The simulation that generates tickets for two customers and displays their info. It also draws the winners for every ticket type.